Sparc: minor fixes

+++ bsp/leon3/pic-sparc-leon3.cpp
@@ -209,7 +209,7 @@ void irq_handler(Mword psr, Mword pc, Mword npc, Mword level)
   // TODO should we manually look for more pending interrupts here?
 //  mgr->c.handle_multi_pending<Chip>(0);
 
-  printf("IRQ #%d handled\n", level);
+  printf("IRQ #%d handled\n", (int)level);
 }
 
 //-------------------------------------------------------------------------------
